server/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
py-mon==2.0.5
python-dotenv==1.0.1
watchdog==5.0.0
Werkzeug==3.0.4
